+++ R/meta_analysis.R
@@ -3,10 +3,11 @@
 #'
 #' @param data A dataframe. It **must** contain columns named `estimate` (effect
 #'   sizes or outcomes)  and `std.error` (corresponding standard errors). These
-#'   two columns will be used for `yi`  and `sei` arguments in `metafor::rma`
-#'   (for parametric analysis) or `metaplus::metaplus` (for robust analysis),
-#'   and for `y` and `SE` arguments in `metaBMA::meta_random` (for Bayesian
-#'   analysis).
+#'   two columns will be used:
+#'   - as `yi`  and `sei` arguments in `metafor::rma` (for **parametric** test)
+#'   or `metaplus::metaplus` (for **robust** test)
+#'   - as `y` and `SE` arguments in `metaBMA::meta_random` (for **Bayesian**
+#'   test).
 #' @inheritParams one_sample_test
 #' @inheritParams metaplus::metaplus
 #' @inheritParams oneway_anova