+++ tests/testthat/test-one_two_sample_dataframes.R
@@ -14,6 +14,7 @@ test_that(
           x = list("brainwt"),
           test.value = list(0.25),
           effsize.type = list("d", "g", "d", "g"),
+          alternative = c("two.sided", "less", "greater", "two.sided"),
           var.equal = list(TRUE, FALSE, TRUE, FALSE),
           conf.level = list(0.89, 0.99, 0.90, 0.50)
         ),