+++ NEWS.md
@@ -1,5 +1,15 @@
 # statsExpressions 1.0.1.9000
 
+MAJOR CHANGES
+
+  - The effect size for independent trimmed means two-sample test has been
+    changed from explanatory measure of effect size to AKP's delta, which is
+    easier to understand and interpret given its a robust cousin of Cohen's
+    *d*.
+
+  - Cohen's *d* and Hedge's *g* use non-pooled standard deviation (cf.
+    https://psyarxiv.com/tu6mp/).
+
 # statsExpressions 1.0.1
 
 BREAKING CHANGES